Refactor main controller and utils
